Changed Dark Mode to Graphical Settings. Added functionality for the start button.
