+++ ConnectFour/app/src/main/java/com/example/connectfour/Board/GameBoard.java
@@ -1,6 +1,7 @@
 package com.example.connectfour.Board;
 
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatDelegate;
 
 import android.os.Bundle;
 
@@ -10,6 +11,12 @@ public class GameBoard extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
+        if(AppCompatDelegate.getDefaultNightMode()==AppCompatDelegate.MODE_NIGHT_YES) {
+            setTheme(R.style.ThemeOverlay_AppCompat_Dark);
+        }
+        else {
+            setTheme(R.style.AppTheme);
+        }
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_game_board);
     }
