+++ ConnectFour/app/src/main/java/com/example/connectfour/NewGameActivity.java
@@ -1,6 +1,7 @@
 package com.example.connectfour;
 
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatDelegate;
 
 import android.os.Bundle;
 
@@ -8,6 +9,12 @@ public class NewGameActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
+        if(AppCompatDelegate.getDefaultNightMode()==AppCompatDelegate.MODE_NIGHT_YES) {
+            setTheme(R.style.ThemeOverlay_AppCompat_Dark);
+        }
+        else {
+            setTheme(R.style.AppTheme);
+        }
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_new_game_screen);
     }
