+++ ConnectFour/app/src/main/java/com/example/connectfour/SettingsActivities/PauseActivity.java
@@ -1,6 +1,7 @@
 package com.example.connectfour.SettingsActivities;
 
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatDelegate;
 
 import android.os.Bundle;
 import android.view.View;
@@ -13,6 +14,12 @@ public class PauseActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
+        if(AppCompatDelegate.getDefaultNightMode()==AppCompatDelegate.MODE_NIGHT_YES) {
+            setTheme(R.style.ThemeOverlay_AppCompat_Dark);
+        }
+        else {
+            setTheme(R.style.AppTheme);
+        }
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_pause);
