+++ ConnectFour/app/src/main/java/com/example/connectfour/StartScreenActivity.java
@@ -12,35 +12,15 @@ import android.widget.Switch;
 
 public class StartScreenActivity extends AppCompatActivity {
 
-    private Switch darkMode;
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
-        if(AppCompatDelegate.getDefaultNightMode()==AppCompatDelegate.MODE_NIGHT_YES) {
+        if (AppCompatDelegate.getDefaultNightMode() == AppCompatDelegate.MODE_NIGHT_YES) {
             setTheme(R.style.ThemeOverlay_AppCompat_Dark);
-        }
-        else {
+        } else {
             setTheme(R.style.AppTheme);
         }
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        darkMode = findViewById(R.id.darkModeSwitch);
-        if(AppCompatDelegate.getDefaultNightMode()==AppCompatDelegate.MODE_NIGHT_YES) {
-            darkMode.setChecked(true);
-        }
-        darkMode.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-            @Override
-            public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
-                if (isChecked) {
-                    AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES);
-                    restartApp();
-                }
-                else {
-                    AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO);
-                    restartApp();
-                }
-            }
-        });
 
         Button start = findViewById(R.id.startGameButton);
         start.setOnClickListener(new View.OnClickListener() {
@@ -50,10 +30,5 @@ public class StartScreenActivity extends AppCompatActivity {
                 startActivity(startgame);
             }
         });
-    }
-    public void restartApp() {
-        Intent intent = new Intent(getApplicationContext(), StartScreenActivity.class);
-        startActivity(intent);
-        finish();
     }
 }
