+++ ConnectFour/app/src/main/res/layout/activity_graphical_settings.xml
@@ -7,14 +7,31 @@
     tools:context=".GraphicalSettings">
 
     <TextView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:firstBaselineToTopHeight="50dp"
-        android:gravity="center_horizontal"
-        android:textSize="25dp"
+        android:id="@+id/textView3"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:firstBaselineToTopHeight="80dp"
+        android:fontFamily="@font/bangers"
         android:text="Graphical Settings"
-        ></TextView>
+        android:textColor="#000000"
+        android:textSize="35dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.065" />
 
+    <Switch
+        android:id="@+id/darkModeSwitch"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Dark Mode"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="@+id/textView3"
+        android:textColor="@color/colorPrimary"
+        />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
